Word to submit has own component
